pyproject.toml
[project]
name = "python-redis-om"
version = "0.1.0"
description = "FastAPI service for Redis OM queries and indexing"
authors = [
    {name = "Sian", email = "anaufal879@gmail.com"}
]
repository = "https://github.com/masan27/python-redis-om"
homepage = "https://github.com/masan27/python-redis-om"
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "redis-om (>=1.1.0,<2.0.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "loguru (>=0.7.3,<1.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"

[tool.poetry.scripts]
migrate = "migrate:run_migrations"

[tool.ruff]
line-length = 88
target-version = "py312"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
